moves methods to utils from main
